requirements.txt
fastapi
uvicorn
pillow
pytesseract
opencv-python
numpy
tensorflow
pandas
matplotlib